Move Grade View code to new activity/class
