+++ app/src/main/java/com/group/a/gradeapp/ViewGradeListActivity.java
@@ -1,8 +1,12 @@
 package com.group.a.gradeapp;
 
+import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import android.os.Bundle;
+import android.view.MenuItem;
 
 public class ViewGradeListActivity extends AppCompatActivity {
 
@@ -10,5 +14,23 @@ public class ViewGradeListActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_view_grade_list);
+
+        ActionBar actionBar = getSupportActionBar();
+        actionBar.setDisplayHomeAsUpEnabled(true);
+
+        RecyclerView recycler_view = findViewById(R.id.gradeview_list);
+
+        RecyclerView.LayoutManager layout_manager = new LinearLayoutManager(this);
+        recycler_view.setLayoutManager(layout_manager);
+
+        java.lang.String[] grades = {};
+
+        GradeViewAdapter grade_adapter = new GradeViewAdapter(grades);
+        recycler_view.setAdapter(grade_adapter);
+    }
+
+    public boolean onOptionsItemSelected(MenuItem item){
+        finish();
+        return true;
     }
 }
