--- app/src/main/res/layout/activity_grades_view.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".GradesView">
-
-    <TextView
-        android:id="@+id/GradesView_title"
-        android:layout_width="match_parent"
-        android:layout_height="45dp"
-        android:gravity="center"
-        android:scaleX="2"
-        android:scaleY="2"
-        android:text="View Grades" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/gradeview_list"
-        android:layout_width="409dp"
-        android:layout_height="match_parent"
-        tools:itemCount="5" />
-</LinearLayout>
