+++ app/src/main/res/layout/activity_view_grade_list.xml
@@ -1,9 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ViewGradeListActivity">
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+xmlns:app="http://schemas.android.com/apk/res-auto"
+xmlns:tools="http://schemas.android.com/tools"
+android:layout_width="match_parent"
+android:layout_height="match_parent"
+android:orientation="vertical"
+tools:context=".ViewGradeListActivity">
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+<TextView
+    android:id="@+id/GradesView_title"
+    android:layout_width="match_parent"
+    android:layout_height="45dp"
+    android:gravity="center"
+    android:scaleX="2"
+    android:scaleY="2"
+    android:text="View Grades" />
+
+<androidx.recyclerview.widget.RecyclerView
+    android:id="@+id/gradeview_list"
+    android:layout_width="409dp"
+    android:layout_height="match_parent"
+    tools:itemCount="5" />
+</LinearLayout>
